moved currentProfile to lib
