+++ app/spondee.tsx
@@ -8,7 +8,7 @@ import * as Speech from "expo-speech";
 import { useEffect, useState } from "react";
 import { StyleSheet, TouchableOpacity, View } from "react-native";
 import SpondeeCards from "../components/spondee/SpondeeCardDefinitions";
-import { userData } from "./currentProfile";
+import { userData } from "../lib/currentProfile";
 
 // Fisher-Yates Shuffle Algorithm
 function shuffleArray<T>(array: T[]): T[] {
