+++ app/starting.tsx
@@ -7,7 +7,7 @@ import {
   TouchableOpacity,
   View,
 } from "react-native";
-import { fetchGuestData } from "./currentProfile";
+import { fetchGuestData } from "../lib/currentProfile";
 
 export default function StartingPointPage() {
   return (
@@ -24,7 +24,6 @@ export default function StartingPointPage() {
         <View style={styles.buttonContainer}>
           <TouchableOpacity
             onPress={() => {
-
               fetchGuestData();
               router.push("/selection");
             }}
